requirements.txt
fastapi
uvicorn
sentence-transformers
PyMuPDF
faiss-cpu
nltk
python-multipart
requests
pydantic
google-generativeai
